+++ test/utils.spec.js
@@ -8,67 +8,111 @@ describe('request-utils', () => {
 
   describe('createRequestOptions()', () => {
 
-    it('should parse hostname, port, path and headers (HTTPS default to 443, default to GET)', (done) => {
+    it('should parse host, port, path and headers (HTTPS default to 443, default to GET)', () => {
 
-      let parsedUrl = url.parse('https://codecloudandrants.io');
+      let parsedUrl = url.parse('https://someurl');
       let requestOptions = createRequestOptions(parsedUrl);
 
-      expect(requestOptions.hostname).to.equal('codecloudandrants.io');
+      expect(requestOptions.host).to.equal('someurl');
       expect(requestOptions.port).to.equal(443);
       expect(requestOptions.path).to.equal('/');
       expect(requestOptions.method).to.equal('GET');
       expect(requestOptions.headers).to.be.undefined;
-
-      done();
     });
 
-    it('should parse should parse hostname, port, path, method and headers (with port in url)', (done) => {
+    it('should parse should parse host, port, path, method and headers (with port in url)', () => {
 
       let options = { method: 'POST', headers: { 'Content-Type': 'application/json' } };
-      let parsedUrl = url.parse('https://codecloudandrants.io:8443/a-path?istrue=true');
+      let parsedUrl = url.parse('https://someurl:8443/a-path?istrue=true');
       let requestOptions = createRequestOptions(parsedUrl, options);
 
-      expect(requestOptions.hostname).to.equal('codecloudandrants.io');
+      expect(requestOptions.host).to.equal('someurl');
       expect(requestOptions.port).to.equal(8443);
       expect(requestOptions.path).to.equal('/a-path?istrue=true');
       expect(requestOptions.method).to.equal('POST');
       expect(requestOptions.headers).to.be.an('Object');
-
-      done();
     });
 
-    it('should parse hostname, port, path and headers (HTTP default to 80)', (done) => {
+    it('should set Content-Length header if a body is in the options without Content-Length', () => {
 
-      let parsedUrl = url.parse('http://codecloudandrants.io');
+      let options = { method: 'POST', body: '{"data":"somedata"}' };
+      let parsedUrl = url.parse('https://someurl:8443/a-path?istrue=true');
+      let requestOptions = createRequestOptions(parsedUrl, options);
+
+      expect(requestOptions.host).to.equal('someurl');
+      expect(requestOptions.port).to.equal(8443);
+      expect(requestOptions.path).to.equal('/a-path?istrue=true');
+      expect(requestOptions.method).to.equal('POST');
+      expect(requestOptions.headers).to.be.an('Object');
+    });
+
+    it('should set Content-Length header if a body is in the options with Content-Length', () => {
+
+      let options = { method: 'POST', body: '{"data":"somedata"}', headers: {'Content-Length': 19 } };
+      let parsedUrl = url.parse('https://someurl:8443/a-path?istrue=true');
+      let requestOptions = createRequestOptions(parsedUrl, options);
+
+      expect(requestOptions.host).to.equal('someurl');
+      expect(requestOptions.port).to.equal(8443);
+      expect(requestOptions.path).to.equal('/a-path?istrue=true');
+      expect(requestOptions.method).to.equal('POST');
+      expect(requestOptions.headers).to.be.an('Object');
+    });
+
+    it('should parse host, port, path and headers (HTTP default to 80)', () => {
+
+      let parsedUrl = url.parse('http://someurl');
       let requestOptions = createRequestOptions(parsedUrl);
 
-      expect(requestOptions.hostname).to.equal('codecloudandrants.io');
+      expect(requestOptions.host).to.equal('someurl');
       expect(requestOptions.port).to.equal(80);
       expect(requestOptions.path).to.equal('/');
       expect(requestOptions.method).to.equal('GET');
       expect(requestOptions.headers).to.be.undefined;
-
-      done();
     });
 
-    it('should parse hostname, port, path and headers (with port in URL)', (done) => {
+    it('should parse host, port, path and headers (with port in URL)', () => {
 
-      let parsedUrl = url.parse('http://codecloudandrants.io:8080');
+      let parsedUrl = url.parse('http://someurl:8080');
       let requestOptions = createRequestOptions(parsedUrl);
 
-      expect(requestOptions.hostname).to.equal('codecloudandrants.io');
+      expect(requestOptions.host).to.equal('someurl');
       expect(requestOptions.port).to.equal(8080);
       expect(requestOptions.path).to.equal('/');
       expect(requestOptions.method).to.equal('GET');
       expect(requestOptions.headers).to.be.undefined;
+    });
 
-      done();
+    it('should add additional ssl/cert options if protocol is https and they are provided in options.certificate', () => {
+
+      let options = { method: 'POST', headers: { 'Content-Type': 'application/json' }, certificate: { ca: 'pathToCert' }};
+      let parsedUrl = url.parse('https://someurl');
+
+      let requestOptions = createRequestOptions(parsedUrl, options);
+
+      expect(requestOptions.host).to.equal('someurl');
+      expect(requestOptions.port).to.equal(443);
+      expect(requestOptions.path).to.equal('/');
+      expect(requestOptions.method).to.equal('POST');
+      expect(requestOptions.headers).to.be.an('Object');
+      expect(requestOptions.ca).to.equal('pathToCert');
+    });
+
+    it('should handle proxy parameter', () => {
+      let options = { method: 'GET', header: { 'Content-Type': 'application/json' }, proxy: 'https://proxy.local:8080' };
+      let parsedUrl = url.parse('https://someurl');
+
+      let requestOptions = createRequestOptions(parsedUrl, options);
+
+      expect(requestOptions.host).to.equal('proxy.local');
+      expect(requestOptions.port).to.equal('8080');
+      expect(requestOptions.headers['Host']).to.equal('https://someurl');
     });
   });
 
   describe('parseResponseBody()', () => {
 
-    it('should try to parse the response as JSON (return JSON object) if no Content-Type is present', (done) => {
+    it('should try to parse the response as JSON (return JSON object) if no Content-Type is present', () => {
       let headers = {};
       let str = '{"data":"some data"}';
       let buf = Buffer.from(str, 'utf8');
@@ -76,11 +120,9 @@ describe('request-utils', () => {
 
       expect(parsedBody).to.be.an('Object');
       expect(parsedBody.data).equal('some data');
-
-      done();
     });
 
-    it('should try to parse the response as JSON (return string) if no Content-Type is present', (done) => {
+    it('should try to parse the response as JSON (return string) if no Content-Type is present', () => {
       let headers = {};
       let str = 'some data';
       let buf = Buffer.from(str, 'utf8');
@@ -88,11 +130,9 @@ describe('request-utils', () => {
 
       expect(parsedBody).to.be.a('string');
       expect(parsedBody).to.equal('some data')
-
-      done();
     });
 
-    it('should try to parse the response as JSON (return JSON object) if Content-Type (application/json) is present', (done) => {
+    it('should try to parse the response as JSON (return JSON object) if Content-Type (application/json) is present', () => {
       let headers = { 'content-type': 'application/json' };
       let str = '{"data":"some data"}';
       let buf = Buffer.from(str, 'utf8');
@@ -100,11 +140,9 @@ describe('request-utils', () => {
 
       expect(parsedBody).to.be.an('Object');
       expect(parsedBody.data).equal('some data');
-
-      done();
     });
 
-    it('should try to parse the response as plain text (return string) if Content-Type (text/html) is present', (done) => {
+    it('should try to parse the response as plain text (return string) if Content-Type (text/html) is present', () => {
       let headers = { 'content-type': 'text/html' };
       let str = 'some data';
       let buf = Buffer.from(str, 'utf8');
@@ -112,11 +150,9 @@ describe('request-utils', () => {
 
       expect(parsedBody).to.be.a('string');
       expect(parsedBody).to.equal('some data')
-
-      done();
     });
 
-    it('should try to parse the response as plain text (return string) if Content-Type (other) is present', (done) => {
+    it('should try to parse the response as plain text (return string) if Content-Type (other) is present', () => {
       let headers = { 'content-type': 'text/plain' };
       let str = 'some data';
       let buf = Buffer.from(str, 'utf8');
@@ -124,26 +160,21 @@ describe('request-utils', () => {
 
       expect(parsedBody).to.be.a('string');
       expect(parsedBody).to.equal('some data')
-
-      done();
     });
   });
 
   describe('isStream()', () => {
 
-    it('should return true if passed in object is a stream', (done) => {
+    it('should return true if passed in object is a stream', () => {
       let rs = fs.createReadStream(__dirname + '/mockfile/read.txt');
       expect(isStream(rs)).to.be.true;
-      done();
     });
 
-    it('shoul return false if passed in object is not a stream', (done) => {
+    it('shoul return false if passed in object is not a stream', () => {
       let obj1 = {}, obj2 = 'string';
 
       expect(isStream(obj1)).to.be.false;
       expect(isStream(obj2)).to.be.false;
-
-      done();
     });
   });
 });